% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jocomo_chisq.R
\name{jocomo.chisq.test}
\alias{jocomo.chisq.test}
\alias{jocomo.chisq.test.default}
\alias{jocomo.chisq.test.formula}
\alias{jocomo.chisq.test.xtabs}
\title{Title}
\usage{
jocomo.chisq.test(...)

\method{jocomo.chisq.test}{default}(
  x,
  y,
  samples,
  models,
  groups,
  correct = F,
  warn = getOption("warn"),
  ...
)

\method{jocomo.chisq.test}{formula}(formula, data = parent.frame(), ...)

\method{jocomo.chisq.test}{xtabs}(xt, data = parent.frame(), correct = FALSE, ...)
}
\arguments{
\item{...}{TODO}

\item{x}{TODO}

\item{y}{TODO}

\item{samples}{TODO}

\item{models}{TODO}

\item{groups}{TODO}

\item{correct}{TODO}

\item{warn}{TODO}

\item{formula}{TODO}

\item{data}{TODO}

\item{xt}{TODO}
}
\value{
TODO

TODO

TODO
}
\description{
Title

Title

Title

Title
}
\examples{
'TODO'
'TODO'
'TODO'
}
