% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jocomo_ratio.R
\name{jocomo.ratio.test.default}
\alias{jocomo.ratio.test.default}
\title{Title}
\usage{
\method{jocomo.ratio.test}{default}(x, y, samples, models, groups)
}
\arguments{
\item{x}{TODO}

\item{y}{TODO}

\item{samples}{TODO}

\item{models}{TODO}

\item{groups}{TODO}
}
\value{
TODO
}
\description{
Title
}
\examples{
'TODO'
}
