% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jocomo_ratio.R
\name{jocomo.ratio.test.formula}
\alias{jocomo.ratio.test.formula}
\title{Title}
\usage{
\method{jocomo.ratio.test}{formula}(formula, data = parent.frame())
}
\arguments{
\item{formula}{TODO}

\item{data}{TODO}
}
\value{
TODO
}
\description{
Title
}
\examples{
'TODO'
}
