% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jocomo_ratio.R
\name{jocomo.ratio.test.xtabs}
\alias{jocomo.ratio.test.xtabs}
\title{Title}
\usage{
\method{jocomo.ratio.test}{xtabs}(xt, data = parent.frame())
}
\arguments{
\item{xt}{TODO}

\item{data}{TODO}
}
\value{
TODO
}
\description{
Title
}
\examples{
'TODO'
}
