% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass_wu.R
\name{multiclass.wu.statistic.default}
\alias{multiclass.wu.statistic.default}
\title{Title}
\usage{
\method{multiclass.wu.statistic}{default}(x, y, correct = F, ...)
}
\arguments{
\item{x}{An object which can be coerced to a `matrix` of size \eqn{p*q} where
\eqn{p} is the number of samples and \eqn{q} is the number of models.
The data should consist of two  or more levels.}

\item{y}{A `vector` of length \eqn{p} where
The data should consist of two or more levels.}

\item{correct}{Add 0.5 to each cell of the 2x2 contingency table to adjust for 0 counts}

\item{...}{Additional arguments passed on to methods.}
}
\description{
Title
}
