% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass_wu.R
\name{multiclass.wu.statistic.xtabs}
\alias{multiclass.wu.statistic.xtabs}
\title{Title}
\usage{
\method{multiclass.wu.statistic}{xtabs}(xt, correct = F, ...)
}
\arguments{
\item{xt}{An `xtabs` object of 3 or more dimensions indicating the
cross-tabulation of model predictions and true labels. Each factor must have
exactly two levels. The first dimension should refer to the true labels
while the remaining dimensions refer to the model predictions.}

\item{correct}{Add 0.5 to each cell of the 2x2 contingency table to adjust for 0 counts}

\item{...}{Additional arguments passed on to methods.}
}
\description{
Title
}
