% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wu.R
\name{wu.statistic.default}
\alias{wu.statistic.default}
\title{Title}
\usage{
\method{wu.statistic}{default}(x, y, correct = F, ...)
}
\arguments{
\item{x}{An object which can be coerced to a `matrix` of size \eqn{p*q} where
\eqn{p} is the number of subjects and \eqn{q} is the number of models.
The data should consist of two ordered levels and is coerced into a logical `matrix`.}

\item{y}{A `vector` of length \eqn{p} where
The data should consist of two ordered levels and is coerced into a logical `vector`.}

\item{correct}{Add 0.5 to each cell of the 2x2 contingency table to adjust for 0 counts}

\item{...}{Additional arguments passed on to methods.}
}
\description{
Title
}
