% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wu.R
\name{wu.statistic}
\alias{wu.statistic}
\title{An implementation of the extended McNemar statistic from Wu 2023 (doi: 10.1080/10543406.2022.2065500)}
\usage{
wu.statistic(x, y, correct = F)
}
\arguments{
\item{x}{a \eqn{p*q} matrix of binary predictions with \eqn{p} subjects as rows and \eqn{q} models as columns.}

\item{y}{a logical vector of length \eqn{p} indicating the positive cases.}

\item{correct}{Add 0.5 to each cell of the 2x2 contingency table to adjust for 0 counts}
}
\description{
An implementation of the extended McNemar statistic from Wu 2023 (doi: 10.1080/10543406.2022.2065500)
}
