% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wu.R
\name{wu.test}
\alias{wu.test}
\alias{wu.test.default}
\alias{wu.test.formula}
\title{An implementation of Wu's test from Wu 2023 (doi: 10.1080/10543406.2022.2065500)}
\usage{
wu.test(x, ...)

\method{wu.test}{default}(x, y, models, subjects, ...)

\method{wu.test}{formula}(formula, data = NULL, ...)
}
\arguments{
\item{x}{a \eqn{p*q} matrix of binary predictions with \eqn{p} subjects as rows and \eqn{q} models as columns.}

\item{...}{further arguments to be passed to or from methods.}

\item{y}{a logical vector of length \eqn{p} indicating the positive cases.}
}
\description{
An implementation of Wu's test from Wu 2023 (doi: 10.1080/10543406.2022.2065500)
}
