% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wu.R
\name{wu.test.default}
\alias{wu.test.default}
\alias{wu.test}
\title{An implementation of Wu's test from Wu 2023 (doi: 10.1080/10543406.2022.2065500)}
\usage{
\method{wu.test}{default}(x, y, threshold = 0.5, ...)

wu.test(x, ...)
}
\arguments{
\item{x}{a \eqn{p*q} matrix of predictions with \eqn{p} subjects as rows and \eqn{q} models as columns.}

\item{y}{a logical vector of length $p$ indicating the positive cases.}

\item{threshold}{If x is not already logical, this threshold is used to determine thresholds for
positive case prediction.  Scores above this threshold are positive cases, while scores equal to
or below the threshold are negative cases. (Default=0.5)}

\item{...}{further arguments to be passed to or from methods.}
}
\description{
An implementation of Wu's test from Wu 2023 (doi: 10.1080/10543406.2022.2065500)
}
