% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wu.R
\name{wu.test}
\alias{wu.test}
\alias{wu.test.default}
\alias{wu.test.formula}
\alias{wu.test.xtabs}
\alias{wu.statistic}
\title{An implementation of Wu's test from Wu 2023 (doi: 10.1080/10543406.2022.2065500)}
\usage{
wu.test(...)

\method{wu.test}{default}(x, y, models, samples, ...)

\method{wu.test}{formula}(formula, data = parent.frame(), ...)

\method{wu.test}{xtabs}(xt, ...)

wu.statistic(...)
}
\arguments{
\item{...}{Additional arguments passed on to methods.}

\item{x}{Can be either a `matrix`, `data.frame`, or `vector` of model predictions.
As a `matrix` or `data.frame`, `x` should be a \eqn{p*q} `matrix` of binary
predictions with \eqn{p} samples as rows and \eqn{q} models as columns. If
`x` is a `vector`, it should have length \eqn{p*q} and both `models` and
`samples` must be specified. The data must be able to be coerced to a `factor`
with two levels. Ignored if a formula is specified.}

\item{y}{If `x` is a `matrix` or `data.frame`, then `y` must be a `vector` of
length \eqn{p} indicating positive and negative cases. If `x` is a `vector`,
then `y` must be a `vector` of length \eqn{p*q}. The data must be able to be
coerced to a `factor` with the same two levels as `x`. Ignored if a formula is
specified.}

\item{models}{A `vector` of length \eqn{p*q} indicating which model the datum
corresponds to. This should have \eqn{q} levels each occurring exactly \eqn{p}
times. Ignored if `x` is a `matrix` or `data.frame` or a formula is specified.}

\item{samples}{A `vector` of length \eqn{p*q} indicating which subject the datum
corresponds to. This should have \eqn{p} levels each occurring exactly \eqn{q}
times. Ignored if `x` is a `matrix` or `data.frame` or a formula is specified.}

\item{formula}{A two-sided formula object describing predictions across
multiple models and samples. Formulas may be specified in either wide, long,
or cross-tabulated format. Refer to 'Details' for more information regarding
formula specification.}

\item{data}{an optional data frame containing the variables named in formula.
By default the variables are taken from the environment from which `wu.test` is
called. While data is optional, the package authors strongly recommend its use.}

\item{xt}{An `xtabs` object of 3 or more dimensions indicating the
cross-tabulation of model predictions and true labels. Each factor must have
exactly two levels. The first dimension should refer to the true labels
while the remaining dimensions refer to the model predictions.}
}
\description{
An implementation of Wu's test from Wu 2023 (doi: 10.1080/10543406.2022.2065500)

Title

Title

An implementation of the extended McNemar statistic from Wu 2023 (doi: 10.1080/10543406.2022.2065500)
}
\details{
Three formats are availabe for formula specification.

Wide format: The term on the left of the ~ operator should
refer to a factor with two levels indicating the true labels for each subject.
Terms should be separated by + operators and refer to predictions from each model.

Long format: The term on the left of the ~ operator should
refer to a factor with two levels indicating the true labels for each subject.
The formula should take the form `x:y|z` where `x` are the model predictions,
`y` describes the model which yielded the prediction, and `z` describes the
subject which the prediction is for.

Cross-tabulated format: The term on the left of the ~ operator should refer to
the frequency, or total counts, for that stratum. Terms should be separated by + operators
and refer to predictions from each model. An additional term, separated from
the others by `||` should be included on the right hand side of the formula.
e.g with 3 models, x1, x2, and x3, and true labels y, the formula should follow
`Freq ~ x1 + x2 + x3 || y`
}
